Add hotchochlate api

+++ src/containers/todos/graphql/Todos.queries.ts
@@ -1,12 +1,7 @@
 import { graphql } from "react-relay";
 
 export const TodosQuery = graphql`
-  query TodosQuery(
-    $before: String
-    $after: String
-    $first: Float
-    $last: Float
-  ) {
+  query TodosQuery($before: String, $after: String, $first: Int, $last: Int) {
     ...TodosFragment
   }
 `;
